+++ r-aks.tf
@@ -6,12 +6,12 @@ resource "azurerm_kubernetes_cluster" "aks" {
   dns_prefix                      = replace(coalesce(var.custom_aks_name, local.aks_name), "/[\\W_]/", "-")
   kubernetes_version              = var.kubernetes_version
   sku_tier                        = var.aks_sku_tier
-  api_server_authorized_ip_ranges = var.enable_private_cluster ? null : var.api_server_authorized_ip_ranges
+  api_server_authorized_ip_ranges = var.private_cluster_enabled ? null : var.api_server_authorized_ip_ranges
   node_resource_group             = var.node_resource_group
   enable_pod_security_policy      = var.enable_pod_security_policy
 
-  private_cluster_enabled = var.enable_private_cluster
-  private_dns_zone_id     = var.enable_private_cluster && var.private_dns_zone_type == "Custom" ? var.private_dns_zone_id : var.private_dns_zone_type
+  private_cluster_enabled = var.private_cluster_enabled
+  private_dns_zone_id     = var.private_cluster_enabled && var.private_dns_zone_type == "Custom" ? var.private_dns_zone_id : var.private_dns_zone_type
 
   default_node_pool {
     name                = local.default_node_pool.name
@@ -29,8 +29,8 @@ resource "azurerm_kubernetes_cluster" "aks" {
   }
 
   identity {
-    type                      = var.enable_private_cluster && var.private_dns_zone_type == "Custom" ? "UserAssigned" : "SystemAssigned"
-    user_assigned_identity_id = var.enable_private_cluster && var.private_dns_zone_type == "Custom" ? azurerm_user_assigned_identity.aks_user_assigned_identity[0].id : null
+    type                      = var.private_cluster_enabled && var.private_dns_zone_type == "Custom" ? "UserAssigned" : "SystemAssigned"
+    user_assigned_identity_id = var.private_cluster_enabled && var.private_dns_zone_type == "Custom" ? azurerm_user_assigned_identity.aks_user_assigned_identity[0].id : null
   }
 
   addon_profile {