+++ tools/agic/helm-agic.tf
@@ -1,5 +1,5 @@
 resource "kubernetes_namespace" "agic" {
-  count = var.enable_agic ? 1 : 0
+  count = var.agic_enabled ? 1 : 0
   metadata {
     name = "system-agic"
     labels = {
@@ -9,14 +9,14 @@ resource "kubernetes_namespace" "agic" {
 }
 
 resource "azurerm_role_assignment" "agic" {
-  count                = var.enable_agic ? 1 : 0
+  count                = var.agic_enabled ? 1 : 0
   principal_id         = var.aks_aad_pod_identity_principal_id
   scope                = azurerm_application_gateway.app_gateway[0].id
   role_definition_name = "Contributor"
 }
 
 resource "azurerm_role_assignment" "agic_rg" {
-  count                = var.enable_agic ? 1 : 0
+  count                = var.agic_enabled ? 1 : 0
   principal_id         = var.aks_aad_pod_identity_principal_id
   scope                = data.azurerm_resource_group.resource_group[0].id
   role_definition_name = "Reader"
@@ -24,7 +24,7 @@ resource "azurerm_role_assignment" "agic_rg" {
 
 
 resource "helm_release" "agic" {
-  count = var.enable_agic ? 1 : 0
+  count = var.agic_enabled ? 1 : 0
   depends_on = [
     azurerm_role_assignment.agic,
     azurerm_role_assignment.agic_rg,