+++ tools/agic/r-logs.tf
@@ -1,5 +1,5 @@
 module "diagnostic_settings_appgw" {
-  count = var.enable_agic ? 1 : 0
+  count = var.agic_enabled ? 1 : 0
 
   source  = "claranet/diagnostic-settings/azurerm"
   version = "4.0.2"